Fix parent link hitbox and toolbar alignment; add Micronaut test resources version

+++ gedbrowserng-frontend/src/app/modules/person/person-parent.component.ts
@@ -5,6 +5,7 @@ import { PersonService } from '../../services';
 import { HasFamily } from '../../interfaces/has-family';
 import { PersonGetter } from './person-getter';
 import { RefreshPerson } from '../../interfaces';
+import { UserService } from '../../services';
 import { RouterLink } from '@angular/router';
 import { MatIconButton } from '@angular/material/button';
 import { MatTooltip } from '@angular/material/tooltip';
@@ -12,19 +13,37 @@ import { MatIcon } from '@angular/material/icon';
 
 @Component({
     selector: 'app-person-parent',
-    template: `
-@if (person) {
-  <span><a class="name"
+    template: `@if (person) {
+  <a class="name"
     [routerLink]="['/' + dataset + '/persons', person.string]">
-    {{ person.indexName }} {{ lifespanYearString() }} [{{ person.string }}]</a>
-  <span class="example-fill-remaining-space"></span>
+    {{ person.indexName }} {{ lifespanYearString() }} [{{ person.string }}]
+  </a>
+}
+@if (person && hasSignedIn()) {
   <span class="hidden">
     <button mat-icon-button matTooltip="Unlink parent" color="warn" (click)="unlink()">
-      <mat-icon matListIcon>link_off</mat-icon></button>
-  </span>
+      <mat-icon matListIcon>link_off</mat-icon>
+    </button>
   </span>
 }`,
-    styles: [],
+    styles: [`
+:host {
+  display: inline-flex;
+  align-items: center;
+  gap: 4px;
+  min-width: 0;
+}
+
+:host .name {
+  display: inline;
+  line-height: 24px;
+  text-decoration: none;
+}
+
+:host:hover .hidden {
+  visibility: visible;
+}
+`],
     imports: [RouterLink, MatIconButton, MatTooltip, MatIcon]
 })
 export class PersonParentComponent extends PersonGetter implements OnInit, OnChanges {
@@ -33,6 +52,7 @@ export class PersonParentComponent extends PersonGetter implements OnInit, OnCha
   @Input() attribute: ApiAttribute;
 
   constructor(@Inject(PersonService) personService: PersonService,
+    @Inject(UserService) private readonly userService: UserService,
     @Inject(NgZone) zone: NgZone,
     @Inject(ChangeDetectorRef) cdr: ChangeDetectorRef) {
     super(personService, zone, cdr);
@@ -53,5 +73,9 @@ export class PersonParentComponent extends PersonGetter implements OnInit, OnCha
 
   refreshPerson() {
     this.parent.refreshPerson();
+  }
+
+  hasSignedIn() {
+    return !!this.userService.currentUser;
   }
 }
